feat : AuthManager의 UserInfo 타입 분리
